+++ .circleci/config.yml
@@ -59,7 +59,6 @@ jobs:
             docker-compose up --build -d loadbalancer
             docker-compose run --rm waitforit -address=tcp://loadbalancer:443 -debug -timeout 120
             docker-compose run --rm waitforit -address=tcp://postgres:5432 -debug
-            docker-compose -f docker-compose.test.yml -f docker-compose.yml run --rm app php bin/console doctrine:migrations:migrate -n
             docker-compose -f docker-compose.test.yml -f docker-compose.yml run --rm app php bin/phpunit --verbose tests
       - run:
           name: Integration Test
