+++ tests/Functional/UploadingCourtOrderTest.php
@@ -5,29 +5,15 @@ namespace App\Tests\Functional;
 
 
 use App\Entity\Order;
-use App\Entity\User;
-use App\Tests\Helpers\OrderTestHelper;
-use Doctrine\Common\DataFixtures\Purger\ORMPurger;
-use Doctrine\ORM\EntityManager;
-use Facebook\WebDriver\Remote\LocalFileDetector;
-use Facebook\WebDriver\Remote\RemoteWebElement;
+use App\Tests\BaseFunctionalTestCase;
 use Facebook\WebDriver\WebDriverBy;
 use Symfony\Component\Panther\Client as PantherClient;
-use Symfony\Component\Panther\PantherTestCase;
 
-class UploadingCourtOrderTest extends PantherTestCase
+class UploadingCourtOrderTest extends BaseFunctionalTestCase
 {
-    const TEST_USER_PASSWORD = 'password123';
-    const TEST_USER_EMAIL = 'test@user.com';
-    const BASIC_AUTH_CREDS = ['PHP_AUTH_USER' => self::TEST_USER_EMAIL, 'PHP_AUTH_PW' => self::TEST_USER_PASSWORD];
-    private $projectDir;
-
     public function setUp(): void
     {
-        self::bootKernel();
-        self::purgeDatabase();
-        self::createTestUser(self::TEST_USER_EMAIL, self::TEST_USER_PASSWORD);
-        $this->projectDir = $this->getService('kernel')->getProjectDir();
+        parent::setUp();
     }
 
     public function testUploadValidWordDoc()
@@ -37,7 +23,7 @@ class UploadingCourtOrderTest extends PantherTestCase
         $caseNumber = $order->getClient()->getCaseNumber();
 
         /** @var PantherClient $client */
-        $client = self::createAuthenticatedClient();
+        $client = self::createAuthenticatedPantherClient();
 
         $client->request('GET', '/case', [], []);
         $crawler = $client->clickLink($caseNumber);
@@ -63,7 +49,7 @@ class UploadingCourtOrderTest extends PantherTestCase
         $caseNumber = $order->getClient()->getCaseNumber();
 
         /** @var PantherClient $client */
-        $client = self::createAuthenticatedClient();
+        $client = self::createAuthenticatedPantherClient();
 
         $client->request('GET', '/case', [], []);
         $crawler = $client->clickLink($caseNumber);
@@ -88,7 +74,7 @@ class UploadingCourtOrderTest extends PantherTestCase
         $caseNumber = $order->getClient()->getCaseNumber();
 
         /** @var PantherClient $client */
-        $client = self::createAuthenticatedClient();
+        $client = self::createAuthenticatedPantherClient();
 
         $client->request('GET', '/case', [], []);
         $crawler = $client->clickLink($caseNumber);
@@ -103,76 +89,5 @@ class UploadingCourtOrderTest extends PantherTestCase
 
         $confirmOrderDetailsForm = $client->getWebDriver()->findElement(WebDriverBy::cssSelector('form[name="confirm_order_details_form"]'));
         self::assertNotNull($confirmOrderDetailsForm->findElement(WebDriverBy::id('confirm_order_details_form_hasAssetsAboveThreshold'))->getText());
-    }
-
-    protected function purgeDatabase()
-    {
-        $purger = new ORMPurger($this->getService('doctrine')->getManager());
-        $purger->purge();
-    }
-
-    protected function createTestUser(string $email, string $password)
-    {
-        $userModel = new User($email);
-        $encodedPassword = $this->getService('security.user_password_encoder.generic')->encodePassword($userModel, $password);
-        $userModel->setPassword($encodedPassword);
-        $this->getEntityManager()->persist($userModel);
-        $this->getEntityManager()->flush();
-    }
-
-    protected function getService($id)
-    {
-        return self::$container->get($id);
-    }
-
-    /**
-     * @return EntityManager
-     */
-    protected function getEntityManager()
-    {
-        return $this->getService('doctrine.orm.entity_manager');
-    }
-
-    /**
-     * @param string $madeAt
-     * @param string $issuedAt
-     * @param string $caseNumber
-     * @param string $orderType
-     * @return \App\Entity\OrderHw
-     * @throws \Doctrine\ORM\ORMException
-     * @throws \Doctrine\ORM\OptimisticLockException
-     */
-    protected function createAndPersistOrder(string $madeAt, string $issuedAt, string $caseNumber, string $orderType)
-    {
-        $order = OrderTestHelper::generateOrder($madeAt, $issuedAt, $caseNumber, $orderType);
-
-        $em = $this->getEntityManager();
-        $em->persist($order);
-        $em->flush();
-        return $order;
-    }
-
-    protected function createAuthenticatedClient()
-    {
-        /** @var PantherClient $client */
-        $client = static::createPantherClient(['external_base_uri' => 'https://loadbalancer']);
-        $client->followRedirects();
-        // @todo find a way to hook into WebDriver cookie and set auth using session
-        // https://symfony.com/doc/current/testing/http_authentication.html https://twitter.com/dunglas/status/1039539719208660992?s=20
-        $client->request('GET', '/login', [], []);
-        $client->submitForm('Sign in', ['_username' => self::TEST_USER_EMAIL, '_password' => self::TEST_USER_PASSWORD]);
-        return $client;
-    }
-
-    /**
-     * @param PantherClient $client
-     * @param string $localFileLocation, the full file path - e.g. '/tests/TestData/filename.png'
-     */
-    protected function uploadDropzoneFile(PantherClient $client, string $localFileLocation)
-    {
-        /** @var RemoteWebElement $fileInput */
-        $fileInput = $client->findElement(WebDriverBy::cssSelector('input[type="file"].dz-hidden-input'));
-        $fileInput->setFileDetector(new LocalFileDetector());
-        $fileInput->sendKeys($this->projectDir . $localFileLocation);
     }
 }
